dev(hansbug): huge update of test and mappings
