+++ test/tree/general/base.py
@@ -5,10 +5,11 @@ from typing import Type
 import numpy as np
 import pytest
 
-from treevalue.tree import func_treelize, TreeValue, raw, mapping, delayed
+from treevalue.tree import func_treelize, TreeValue, raw, mapping, delayed, FastTreeValue
+from ..tree.base import get_treevalue_test
 
 
-def get_tree_test(tree_value_clazz: Type[TreeValue]):
+def get_fasttreevalue_test(treevalue_class: Type[FastTreeValue]):
     class Container:
         def __init__(self, value):
             self.__value = value
@@ -36,18 +37,18 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
 
     # noinspection DuplicatedCode,PyMethodMayBeStatic
     @pytest.mark.unittest
-    class _TestClass:
+    class _TestClass(get_treevalue_test(treevalue_class)):
         def test_basic_methods(self):
-            t3 = tree_value_clazz({'a': 14, 'b': 26, 'x': {'c': 38, 'd': 11}})
+            t3 = treevalue_class({'a': 14, 'b': 26, 'x': {'c': 38, 'd': 11}})
             assert t3.json() == {'a': 14, 'b': 26, 'x': {'c': 38, 'd': 11}}
 
             t5 = t3.x.clone()
-            assert t5 == tree_value_clazz({'c': 38, 'd': 11})
+            assert t5 == treevalue_class({'c': 38, 'd': 11})
 
         def test_numeric_add(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (2 + t1 + t2) == tree_value_clazz({'a': 14, 'b': 26, 'x': {'c': 38, 'd': 11}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (2 + t1 + t2) == treevalue_class({'a': 14, 'b': 26, 'x': {'c': 38, 'd': 11}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -58,9 +59,9 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_sub(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (2 - t1 - t2) == tree_value_clazz({'a': -10, 'b': -22, 'x': {'c': -34, 'd': -7}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (2 - t1 - t2) == treevalue_class({'a': -10, 'b': -22, 'x': {'c': -34, 'd': -7}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -71,18 +72,18 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_mul(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (2 * t1 * t2) == tree_value_clazz({'a': 22, 'b': 88, 'x': {'c': 198, 'd': 40}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (2 * t1 * t2) == treevalue_class({'a': 22, 'b': 88, 'x': {'c': 198, 'd': 40}})
 
             i, j = 1, 1.0
             assert i.__mul__(j) == NotImplemented
             assert j.__mul__(i) == 1.0
             assert i * j == j.__mul__(i)
 
-            t3 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5.2}})
-            assert (t1 * t3) == tree_value_clazz({'a': 11, 'b': 44, 'x': {'c': 99, 'd': 20.8}})
-            assert (t3 * t1) == tree_value_clazz({'a': 11, 'b': 44, 'x': {'c': 99, 'd': 20.8}})
+            t3 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5.2}})
+            assert (t1 * t3) == treevalue_class({'a': 11, 'b': 44, 'x': {'c': 99, 'd': 20.8}})
+            assert (t3 * t1) == treevalue_class({'a': 11, 'b': 44, 'x': {'c': 99, 'd': 20.8}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -93,7 +94,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_matmul(self):
-            t1 = tree_value_clazz({
+            t1 = treevalue_class({
                 'a': np.array([[1, 2], [3, 4]]),
                 'b': np.array([[2, 3], [4, 5]]),
                 'x': {
@@ -101,7 +102,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                     'd': np.array([[4, 5], [6, 7]]),
                 }
             })
-            t2 = tree_value_clazz({
+            t2 = treevalue_class({
                 'a': np.array([[4, 5], [6, 7]]),
                 'b': np.array([[3, 4], [5, 6]]),
                 'x': {
@@ -110,15 +111,15 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 }
             })
 
-            tnp_array_equal = func_treelize(return_type=tree_value_clazz)(np.array_equal)
-            assert tnp_array_equal((t1 @ t2), tree_value_clazz({
+            tnp_array_equal = func_treelize(return_type=treevalue_class)(np.array_equal)
+            assert tnp_array_equal((t1 @ t2), treevalue_class({
                 'a': np.array([[1, 2], [3, 4]]) @ np.array([[4, 5], [6, 7]]),
                 'b': np.array([[2, 3], [4, 5]]) @ np.array([[3, 4], [5, 6]]),
                 'x': {
                     'c': np.array([[3, 4], [5, 6]]) @ np.array([[2, 3], [4, 5]]),
                     'd': np.array([[4, 5], [6, 7]]) @ np.array([[1, 2], [3, 4]]),
                 }
-            })) == tree_value_clazz({
+            })) == treevalue_class({
                 'a': True,
                 'b': True,
                 'x': {
@@ -127,7 +128,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 }
             })
             assert tnp_array_equal(
-                (t2.__rmatmul__(np.array([[1, 2], [3, 4]]))), tree_value_clazz({
+                (t2.__rmatmul__(np.array([[1, 2], [3, 4]]))), treevalue_class({
                     'a': np.array([[1, 2], [3, 4]]) @ np.array([[4, 5], [6, 7]]),
                     'b': np.array([[1, 2], [3, 4]]) @ np.array([[3, 4], [5, 6]]),
                     'x': {
@@ -135,7 +136,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                         'd': np.array([[1, 2], [3, 4]]) @ np.array([[1, 2], [3, 4]]),
                     }
                 })
-            ) == tree_value_clazz({
+            ) == treevalue_class({
                 'a': True,
                 'b': True,
                 'x': {
@@ -180,10 +181,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(tt1.x._detach()) == original_id_x
 
         def test_numeric_floordiv(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t2 / t1) == tree_value_clazz({'a': 11.0, 'b': 11.0, 'x': {'c': 11.0, 'd': 1.25}})
-            assert (6 / t1) == tree_value_clazz({'a': 6.0, 'b': 3.0, 'x': {'c': 2.0, 'd': 1.5}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t2 / t1) == treevalue_class({'a': 11.0, 'b': 11.0, 'x': {'c': 11.0, 'd': 1.25}})
+            assert (6 / t1) == treevalue_class({'a': 6.0, 'b': 3.0, 'x': {'c': 2.0, 'd': 1.5}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -194,10 +195,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_truediv(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t2 // t1) == tree_value_clazz({'a': 11, 'b': 11, 'x': {'c': 11, 'd': 1}})
-            assert (6 // t1) == tree_value_clazz({'a': 6, 'b': 3, 'x': {'c': 2, 'd': 1}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t2 // t1) == treevalue_class({'a': 11, 'b': 11, 'x': {'c': 11, 'd': 1}})
+            assert (6 // t1) == treevalue_class({'a': 6, 'b': 3, 'x': {'c': 2, 'd': 1}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -208,10 +209,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_mod(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t2 % t1) == tree_value_clazz({'a': 0, 'b': 0, 'x': {'c': 0, 'd': 1}})
-            assert (6 % t1) == tree_value_clazz({'a': 0, 'b': 0, 'x': {'c': 0, 'd': 2}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t2 % t1) == treevalue_class({'a': 0, 'b': 0, 'x': {'c': 0, 'd': 1}})
+            assert (6 % t1) == treevalue_class({'a': 0, 'b': 0, 'x': {'c': 0, 'd': 2}})
 
             original_id = id(t2._detach())
             original_id_x = id(t2.x._detach())
@@ -222,9 +223,9 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t2.x._detach()) == original_id_x
 
         def test_numeric_power(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            assert (t1 ** t1) == tree_value_clazz({'a': 1, 'b': 4, 'x': {'c': 27, 'd': 256}})
-            assert (2 ** t1) == tree_value_clazz({'a': 2, 'b': 4, 'x': {'c': 8, 'd': 16}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            assert (t1 ** t1) == treevalue_class({'a': 1, 'b': 4, 'x': {'c': 27, 'd': 256}})
+            assert (2 ** t1) == treevalue_class({'a': 2, 'b': 4, 'x': {'c': 8, 'd': 16}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -235,10 +236,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_and(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t1 & t2) == tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 1, 'd': 4}})
-            assert (7 & t2) == tree_value_clazz({'a': 3, 'b': 6, 'x': {'c': 1, 'd': 5}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t1 & t2) == treevalue_class({'a': 1, 'b': 2, 'x': {'c': 1, 'd': 4}})
+            assert (7 & t2) == treevalue_class({'a': 3, 'b': 6, 'x': {'c': 1, 'd': 5}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -249,10 +250,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_or(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t1 | t2) == tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 35, 'd': 5}})
-            assert (7 | t2) == tree_value_clazz({'a': 15, 'b': 23, 'x': {'c': 39, 'd': 7}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t1 | t2) == treevalue_class({'a': 11, 'b': 22, 'x': {'c': 35, 'd': 5}})
+            assert (7 | t2) == treevalue_class({'a': 15, 'b': 23, 'x': {'c': 39, 'd': 7}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -263,10 +264,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_xor(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t1 ^ t2) == tree_value_clazz({'a': 10, 'b': 20, 'x': {'c': 34, 'd': 1}})
-            assert (7 ^ t2) == tree_value_clazz({'a': 12, 'b': 17, 'x': {'c': 38, 'd': 2}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t1 ^ t2) == treevalue_class({'a': 10, 'b': 20, 'x': {'c': 34, 'd': 1}})
+            assert (7 ^ t2) == treevalue_class({'a': 12, 'b': 17, 'x': {'c': 38, 'd': 2}})
 
             original_id = id(t1._detach())
             original_id_x = id(t1.x._detach())
@@ -277,10 +278,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t1.x._detach()) == original_id_x
 
         def test_numeric_lshift(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t2 << t1) == tree_value_clazz({'a': 22, 'b': 88, 'x': {'c': 264, 'd': 80}})
-            assert (3 << t1) == tree_value_clazz({'a': 6, 'b': 12, 'x': {'c': 24, 'd': 48}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t2 << t1) == treevalue_class({'a': 22, 'b': 88, 'x': {'c': 264, 'd': 80}})
+            assert (3 << t1) == treevalue_class({'a': 6, 'b': 12, 'x': {'c': 24, 'd': 48}})
 
             original_id = id(t2._detach())
             original_id_x = id(t2.x._detach())
@@ -291,10 +292,10 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t2.x._detach()) == original_id_x
 
         def test_numeric_rshift(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
-            assert (t2 >> t1) == tree_value_clazz({'a': 5, 'b': 5, 'x': {'c': 4, 'd': 0}})
-            assert (32 >> t1) == tree_value_clazz({'a': 16, 'b': 8, 'x': {'c': 4, 'd': 2}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 22, 'x': {'c': 33, 'd': 5}})
+            assert (t2 >> t1) == treevalue_class({'a': 5, 'b': 5, 'x': {'c': 4, 'd': 0}})
+            assert (32 >> t1) == treevalue_class({'a': 16, 'b': 8, 'x': {'c': 4, 'd': 2}})
 
             original_id = id(t2._detach())
             original_id_x = id(t2.x._detach())
@@ -305,30 +306,30 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert id(t2.x._detach()) == original_id_x
 
         def test_numeric_pos(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            assert +t1 == tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            assert +t1 == treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
 
         def test_numeric_neg(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            assert -t1 == tree_value_clazz({'a': -1, 'b': -2, 'x': {'c': -3, 'd': -4}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            assert -t1 == treevalue_class({'a': -1, 'b': -2, 'x': {'c': -3, 'd': -4}})
 
         def test_numeric_invert(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            assert ~t1 == tree_value_clazz({'a': -2, 'b': -3, 'x': {'c': -4, 'd': -5}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            assert ~t1 == treevalue_class({'a': -2, 'b': -3, 'x': {'c': -4, 'd': -5}})
 
         def test_getitem(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': [2, 3, 5, 7], 'b': [11, 13, 17, 19],
-                                   'x': {'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': [2, 3, 5, 7], 'b': [11, 13, 17, 19],
+                                  'x': {'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]}})
 
-            assert t2[t1 - 1] == tree_value_clazz({'a': 2, 'b': 13, 'x': {'c': 31, 'd': 53}})
+            assert t2[t1 - 1] == treevalue_class({'a': 2, 'b': 13, 'x': {'c': 31, 'd': 53}})
             assert t2['a'] == [2, 3, 5, 7]
-            assert t2['x'] == tree_value_clazz({'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]})
+            assert t2['x'] == treevalue_class({'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]})
             assert t2['x']['c'] == [23, 29, 31, 37]
 
             m1 = np.random.random((3, 4))
             m2 = np.random.random((3, 5))
-            a = tree_value_clazz({'a': m1, 'b': m2})
+            a = treevalue_class({'a': m1, 'b': m2})
             assert np.allclose(a['a'], m1)
             assert np.allclose(a['b'], m2)
 
@@ -340,68 +341,68 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert np.allclose(a_35['a'], m1[:, 3:5])
             assert np.allclose(a_35['b'], m2[:, 3:5])
 
-            b = tree_value_clazz({'': m1, '1': m2})
+            b = treevalue_class({'': m1, '1': m2})
             assert np.allclose(b[''], m1)
             assert np.allclose(b['1'], m2)
 
-            t3 = tree_value_clazz({
+            t3 = treevalue_class({
                 'a': raw({'a': 1, 'y': 2}),
                 'c': {'x': raw({'a': 3, 'y': 4})},
             })
             assert t3['a'] == {'a': 1, 'y': 2}
-            assert t3['c'] == tree_value_clazz({'x': raw({'a': 3, 'y': 4})})
+            assert t3['c'] == treevalue_class({'x': raw({'a': 3, 'y': 4})})
             assert t3['c']['x'] == {'a': 3, 'y': 4}
             with pytest.raises(KeyError):
                 _ = t3['y']
 
-            assert t3[['a']] == tree_value_clazz({'a': 1, 'c': {'x': 3}})
-            assert t3[['y']] == tree_value_clazz({'a': 2, 'c': {'x': 4}})
+            assert t3[['a']] == treevalue_class({'a': 1, 'c': {'x': 3}})
+            assert t3[['y']] == treevalue_class({'a': 2, 'c': {'x': 4}})
 
         def test_setitem(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': [2, 3, 5, 7], 'b': [11, 13, 17, 19],
-                                   'x': {'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': [2, 3, 5, 7], 'b': [11, 13, 17, 19],
+                                  'x': {'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]}})
 
             t2[t1 - 1] = t1
-            assert t2 == tree_value_clazz({'a': [1, 3, 5, 7], 'b': [11, 2, 17, 19],
-                                           'x': {'c': [23, 29, 3, 37], 'd': [41, 43, 47, 4]}})
+            assert t2 == treevalue_class({'a': [1, 3, 5, 7], 'b': [11, 2, 17, 19],
+                                          'x': {'c': [23, 29, 3, 37], 'd': [41, 43, 47, 4]}})
 
-            t3 = tree_value_clazz({
+            t3 = treevalue_class({
                 'a': raw({'a': 1, 'y': 2}),
                 'c': {'x': raw({'a': 3, 'y': 4})},
             })
             t3['a'] = {'a': 11, 'y': 22}
-            assert t3 == tree_value_clazz({
+            assert t3 == treevalue_class({
                 'a': raw({'a': 11, 'y': 22}),
                 'c': {'x': raw({'a': 3, 'y': 4})},
             })
             t3[['a']] = 33
-            assert t3 == tree_value_clazz({
+            assert t3 == treevalue_class({
                 'a': raw({'a': 33, 'y': 22}),
                 'c': {'x': raw({'a': 33, 'y': 4})},
             })
             t3[['y']] = 55
-            assert t3 == tree_value_clazz({
+            assert t3 == treevalue_class({
                 'a': raw({'a': 33, 'y': 55}),
                 'c': {'x': raw({'a': 33, 'y': 55})},
             })
 
         def test_delitem(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': [2, 3, 5, 7], 'b': [11, 13, 17, 19],
-                                   'x': {'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': [2, 3, 5, 7], 'b': [11, 13, 17, 19],
+                                  'x': {'c': [23, 29, 31, 37], 'd': [41, 43, 47, 53]}})
 
             del t2[t1 - 1]
-            assert t2 == tree_value_clazz({'a': [3, 5, 7], 'b': [11, 17, 19],
-                                           'x': {'c': [23, 29, 37], 'd': [41, 43, 47]}})
+            assert t2 == treevalue_class({'a': [3, 5, 7], 'b': [11, 17, 19],
+                                          'x': {'c': [23, 29, 37], 'd': [41, 43, 47]}})
 
-            t3 = tree_value_clazz({
+            t3 = treevalue_class({
                 'a': raw({'a': 1, 'y': 2}),
                 'c': {'x': raw({'a': 3, 'y': 4})},
                 'g': 2,
             })
             del t3['g']
-            assert t3 == tree_value_clazz({
+            assert t3 == treevalue_class({
                 'a': raw({'a': 1, 'y': 2}),
                 'c': {'x': raw({'a': 3, 'y': 4})},
             })
@@ -410,41 +411,41 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 del t3[['g']]
 
             del t3[['a']]
-            assert t3 == tree_value_clazz({
+            assert t3 == treevalue_class({
                 'a': raw({'y': 2}),
                 'c': {'x': raw({'y': 4})},
             })
 
             del t3['a']
-            assert t3 == tree_value_clazz({
+            assert t3 == treevalue_class({
                 'c': {'x': raw({'y': 4})},
             })
 
         def test_attr(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = func_treelize(return_type=tree_value_clazz)(Container)(t1)
-            assert t2 == tree_value_clazz(
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = func_treelize(return_type=treevalue_class)(Container)(t1)
+            assert t2 == treevalue_class(
                 {'a': Container(1), 'b': Container(2), 'x': {'c': Container(3), 'd': Container(4)}})
             assert t2.value == t1
 
             assert t1.a == 1
-            assert t1.x == tree_value_clazz({'c': 3, 'd': 4})
+            assert t1.x == treevalue_class({'c': 3, 'd': 4})
 
             m1 = np.random.random((3, 4))
             m2 = np.random.random((3, 5))
-            a = tree_value_clazz({'a': m1, 'b': m2})
+            a = treevalue_class({'a': m1, 'b': m2})
             assert np.allclose(a.a, m1)
             assert np.allclose(a.b, m2)
-            assert a.shape == tree_value_clazz({
+            assert a.shape == treevalue_class({
                 'a': (3, 4), 'b': (3, 5),
             })
 
         def test_call(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = func_treelize(return_type=tree_value_clazz)(Container)(t1)
-            assert t2.add(10) == tree_value_clazz({'a': 11, 'b': 12, 'x': {'c': 13, 'd': 14}})
-            assert t2.add(x=10) == tree_value_clazz({'a': 11, 'b': 12, 'x': {'c': 13, 'd': 14}})
-            assert t2.add(t1) == tree_value_clazz({'a': 2, 'b': 4, 'x': {'c': 6, 'd': 8}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = func_treelize(return_type=treevalue_class)(Container)(t1)
+            assert t2.add(10) == treevalue_class({'a': 11, 'b': 12, 'x': {'c': 13, 'd': 14}})
+            assert t2.add(x=10) == treevalue_class({'a': 11, 'b': 12, 'x': {'c': 13, 'd': 14}})
+            assert t2.add(t1) == treevalue_class({'a': 2, 'b': 4, 'x': {'c': 6, 'd': 8}})
 
         def test_map(self):
             cnt = 0
@@ -454,14 +455,14 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 cnt += 1
                 return x + 2
 
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
             assert cnt == 0
             t2 = t1.map(f)
             assert cnt == 4
-            assert t2 == tree_value_clazz({'a': 3, 'b': 4, 'x': {'c': 5, 'd': 6}})
+            assert t2 == treevalue_class({'a': 3, 'b': 4, 'x': {'c': 5, 'd': 6}})
 
             cnt = 0
-            t3 = tree_value_clazz({
+            t3 = treevalue_class({
                 'a': delayed(lambda: t1.a),
                 'b': delayed(lambda: t1.b),
                 'x': delayed(lambda: t1.x),
@@ -474,48 +475,48 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert t4.a == 3
             assert cnt == 1
 
-            assert t4 == tree_value_clazz({'a': 3, 'b': 4, 'x': {'c': 5, 'd': 6}})
+            assert t4 == treevalue_class({'a': 3, 'b': 4, 'x': {'c': 5, 'd': 6}})
             assert cnt == 4
 
             assert t4.a == 3
             assert cnt == 4
 
         def test_type(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
             assert t1.type(TreeValue) == TreeValue({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            assert t1.type(TreeValue) != tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            assert t1.type(TreeValue) != treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
 
         def test_filter(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            assert t1.filter(lambda x: x % 2 == 1) == tree_value_clazz({'a': 1, 'x': {'c': 3}})
-            assert t1.filter(lambda x: x < 3) == tree_value_clazz({'a': 1, 'b': 2, })
-            assert t1.filter(lambda x: x < 3, False) == tree_value_clazz({'a': 1, 'b': 2, 'x': {}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            assert t1.filter(lambda x: x % 2 == 1) == treevalue_class({'a': 1, 'x': {'c': 3}})
+            assert t1.filter(lambda x: x < 3) == treevalue_class({'a': 1, 'b': 2, })
+            assert t1.filter(lambda x: x < 3, False) == treevalue_class({'a': 1, 'b': 2, 'x': {}})
 
         def test_mask(self):
-            t1 = tree_value_clazz({'a': 13, 'b': 27, 'x': {'c': 39, 'd': 45}})
-            t2 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t3 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 7, 'd': 4}})
+            t1 = treevalue_class({'a': 13, 'b': 27, 'x': {'c': 39, 'd': 45}})
+            t2 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t3 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 7, 'd': 4}})
 
             mask1 = t2.map(lambda x: (lambda v: v % x == 0))(t1)
-            assert t1.mask(mask1) == tree_value_clazz({'a': 13, 'x': {'c': 39}})
+            assert t1.mask(mask1) == treevalue_class({'a': 13, 'x': {'c': 39}})
 
             mask2 = t3.map(lambda x: (lambda v: v % x == 0))(t1)
-            assert t1.mask(mask2) == tree_value_clazz({'a': 13})
-            assert t1.mask(mask2, False) == tree_value_clazz({'a': 13, 'x': {}})
+            assert t1.mask(mask2) == treevalue_class({'a': 13})
+            assert t1.mask(mask2, False) == treevalue_class({'a': 13, 'x': {}})
 
         def test_reduce(self):
-            t1 = tree_value_clazz({'a': 13, 'b': 27, 'x': {'c': 39, 'd': 45}})
-            t2 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t1 = treevalue_class({'a': 13, 'b': 27, 'x': {'c': 39, 'd': 45}})
+            t2 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
 
             assert t1.reduce(lambda **kwargs: sum(kwargs.values())) == 124
             assert t2.reduce(lambda **kwargs: reduce(__mul__, kwargs.values())) == 24
 
         def test_union(self):
-            t1 = tree_value_clazz({'a': 13, 'b': 27, 'x': {'c': 39, 'd': 45}})
-            t2 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t3 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 7, 'd': 4}})
+            t1 = treevalue_class({'a': 13, 'b': 27, 'x': {'c': 39, 'd': 45}})
+            t2 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t3 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 7, 'd': 4}})
 
-            assert tree_value_clazz.union(t1, t2, t3) == tree_value_clazz({
+            assert treevalue_class.union(t1, t2, t3) == treevalue_class({
                 'a': (13, 1, 1),
                 'b': (27, 2, 2),
                 'x': {
@@ -524,7 +525,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 }
             })
 
-            assert tree_value_clazz.union(t1, t2, t3, return_type=TreeValue) == TreeValue({
+            assert treevalue_class.union(t1, t2, t3, return_type=TreeValue) == TreeValue({
                 'a': (13, 1, 1),
                 'b': (27, 2, 2),
                 'x': {
@@ -546,11 +547,11 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 'k': '233'
             }
 
-            assert tree_value_clazz.subside(data) == tree_value_clazz({
+            assert treevalue_class.subside(data) == treevalue_class({
                 'a': raw({'a': 1, 'k': '233', 'x': {'c': 3, 'd': [5, 7]}}),
                 'b': raw({'a': 2, 'k': '233', 'x': {'c': 4, 'd': [6, 8]}}),
             })
-            assert tree_value_clazz.subside(data, return_type=TreeValue) == TreeValue({
+            assert treevalue_class.subside(data, return_type=TreeValue) == TreeValue({
                 'a': raw({'a': 1, 'k': '233', 'x': {'c': 3, 'd': [5, 7]}}),
                 'b': raw({'a': 2, 'k': '233', 'x': {'c': 4, 'd': [6, 8]}}),
             })
@@ -568,21 +569,21 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 'k': '233'
             }
 
-            t1 = tree_value_clazz.subside(data)
+            t1 = treevalue_class.subside(data)
             assert t1.rise() == {
-                'a': tree_value_clazz({'a': 1, 'b': 2}),
+                'a': treevalue_class({'a': 1, 'b': 2}),
                 'x': {
-                    'c': tree_value_clazz({'a': 3, 'b': 4}),
+                    'c': treevalue_class({'a': 3, 'b': 4}),
                     'd': [
-                        tree_value_clazz({'a': 5, 'b': 6}),
-                        tree_value_clazz({'a': 7, 'b': 8}),
+                        treevalue_class({'a': 5, 'b': 6}),
+                        treevalue_class({'a': 7, 'b': 8}),
                     ]
                 },
-                'k': tree_value_clazz({'a': '233', 'b': '233'}),
+                'k': treevalue_class({'a': '233', 'b': '233'}),
             }
 
         def test_deep_clone(self):
-            t = tree_value_clazz({
+            t = treevalue_class({
                 'a': raw({'a': 1, 'b': 2}),
                 'b': raw({'a': 3, 'b': 4}),
                 'x': {
@@ -606,7 +607,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert t2.x.d is not t.x.d
 
         def test_graph(self):
-            t = tree_value_clazz({
+            t = treevalue_class({
                 'a': [4, 3, 2, 1],
                 'b': np.array([[5, 6], [7, 8]]),
                 'x': {
@@ -619,7 +620,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert len(graph.source) <= 2290
 
         def test_graphics(self):
-            t = tree_value_clazz({
+            t = treevalue_class({
                 'a': [4, 3, 2, 1],
                 'b': np.array([[5, 6], [7, 8]]),
                 'x': {
@@ -628,7 +629,7 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                     'e': np.array([[1, 2], [3, 4]])
                 },
             })
-            t1 = tree_value_clazz({
+            t1 = treevalue_class({
                 'aa': t.a,
                 'bb': np.array([[5, 6], [7, 8]]),
                 'xx': {
@@ -638,45 +639,45 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
                 },
             })
 
-            graph_1 = tree_value_clazz.graphics(
+            graph_1 = treevalue_class.graphics(
                 (t, 't'), (t1, 't1'),
-                (tree_value_clazz({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
+                (treevalue_class({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
                 dup_value=(np.ndarray, list),
                 title="This is a demo of 2 trees with dup value.",
                 cfg={'bgcolor': '#ffffffff'},
             )
             assert len(graph_1.source) <= 4960
 
-            graph_2 = tree_value_clazz.graphics(
+            graph_2 = treevalue_class.graphics(
                 (t, 't'), (t1, 't1'),
-                (tree_value_clazz({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
+                (treevalue_class({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
                 dup_value=False,
                 title="This is a demo of 2 trees with dup value.",
                 cfg={'bgcolor': '#ffffffff'},
             )
             assert len(graph_2.source) <= 5480
 
-            graph_3 = tree_value_clazz.graphics(
+            graph_3 = treevalue_class.graphics(
                 (t, 't'), (t1, 't1'),
-                (tree_value_clazz({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
+                (treevalue_class({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
                 dup_value=lambda x: id(x),
                 title="This is a demo of 2 trees with dup value.",
                 cfg={'bgcolor': '#ffffffff'},
             )
             assert len(graph_3.source) <= 4760
 
-            graph_4 = tree_value_clazz.graphics(
+            graph_4 = treevalue_class.graphics(
                 (t, 't'), (t1, 't1'),
-                (tree_value_clazz({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
+                (treevalue_class({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
                 dup_value=lambda x: type(x).__name__,
                 title="This is a demo of 2 trees with dup value.",
                 cfg={'bgcolor': '#ffffffff'},
             )
             assert len(graph_4.source) <= 3780
 
-            graph_6 = tree_value_clazz.graphics(
+            graph_6 = treevalue_class.graphics(
                 (t, 't'), (t1, 't1'),
-                (tree_value_clazz({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
+                (treevalue_class({'a': t, 'b': t1, 'c': [1, 2], 'd': t1.xx}), 't2'),
                 dup_value=True,
                 title="This is a demo of 2 trees with dup value.",
                 cfg={'bgcolor': '#ffffffff'},
@@ -684,18 +685,18 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
             assert len(graph_6.source) <= 4760
 
         def test_func(self):
-            t1 = tree_value_clazz({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
-            t2 = tree_value_clazz({'a': 11, 'b': 20, 'x': {'c': 33, 'd': 48}})
+            t1 = treevalue_class({'a': 1, 'b': 2, 'x': {'c': 3, 'd': 4}})
+            t2 = treevalue_class({'a': 11, 'b': 20, 'x': {'c': 33, 'd': 48}})
 
-            @tree_value_clazz.func()
+            @treevalue_class.func()
             def ssum(x, y):
                 return x + y
 
-            assert ssum(t1, t2) == tree_value_clazz({'a': 12, 'b': 22, 'x': {'c': 36, 'd': 52}})
+            assert ssum(t1, t2) == treevalue_class({'a': 12, 'b': 22, 'x': {'c': 36, 'd': 52}})
 
             cnt_1 = 0
 
-            @tree_value_clazz.func(delayed=True)
+            @treevalue_class.func(delayed=True)
             def ssumx(x, y):
                 nonlocal cnt_1
                 cnt_1 += 1
@@ -707,20 +708,83 @@ def get_tree_test(tree_value_clazz: Type[TreeValue]):
 
             assert t3.a == 12
             assert cnt_1 == 1
-            assert t3.x == tree_value_clazz({'c': 36, 'd': 52})
+            assert t3.x == treevalue_class({'c': 36, 'd': 52})
             assert cnt_1 == 3
-            assert t3 == tree_value_clazz({'a': 12, 'b': 22, 'x': {'c': 36, 'd': 52}})
+            assert t3 == treevalue_class({'a': 12, 'b': 22, 'x': {'c': 36, 'd': 52}})
             assert cnt_1 == 4
 
         def test_walk(self):
-            tv1 = tree_value_clazz({'a': 1, 'b': 'dks', 'c': {'x': 2, 'y': 3}})
+            tv1 = treevalue_class({'a': 1, 'b': 'dks', 'c': {'x': 2, 'y': 3}})
             assert dict(tv1.walk()) == {
-                (): tree_value_clazz({'a': 1, 'b': 'dks', 'c': {'x': 2, 'y': 3}}),
+                (): treevalue_class({'a': 1, 'b': 'dks', 'c': {'x': 2, 'y': 3}}),
                 ('a',): 1,
                 ('b',): 'dks',
-                ('c',): tree_value_clazz({'x': 2, 'y': 3}),
+                ('c',): treevalue_class({'x': 2, 'y': 3}),
                 ('c', 'x'): 2,
                 ('c', 'y'): 3,
             }
+
+        def test_tree_value_operate_with_item(self):
+            tv1 = treevalue_class({'a': 1, 'b': 2, 'c': {'x': 2, 'y': 3}})
+            tv2 = treevalue_class(tv1)
+            tv3 = treevalue_class({'a': tv1, 'b': tv2, 'c': tv1})
+            tv4 = treevalue_class({'a': raw({'a': 1, 'y': 2}), 'c': {'x': raw({'a': 3, 'y': 4}), }})
+
+            tv1['a'] = 3
+            assert tv1.a == 3
+            assert tv2.a == 3
+            assert tv3.a.a == 3
+            assert tv3.b.a == 3
+            assert tv3.c.a == 3
+
+            assert tv4['a'] == {'a': 1, 'y': 2}
+            assert tv4['c'] == treevalue_class({'x': raw({'a': 3, 'y': 4})})
+            with pytest.raises(KeyError):
+                _ = tv4['y']
+            with pytest.raises(KeyError):
+                _ = tv4[['c']]
+
+            tv1['f'] = 333
+            assert tv1.f == 333
+            assert tv1['f'] == 333
+            assert 'f' in tv1
+
+            with pytest.raises(TypeError):
+                tv1[0] = 3
+            with pytest.raises(TypeError):
+                tv1[['c']] = 3
+
+            del tv1['b']
+            assert 'b' not in tv1
+            assert 'b' not in tv2
+            assert 'b' not in tv3.a
+            assert 'b' not in tv3.b
+            assert 'b' not in tv3.c
+
+            with pytest.raises(KeyError):
+                del tv1['g']
+            with pytest.raises(TypeError):
+                del tv1[['c']]
+            with pytest.raises(TypeError):
+                del tv1[0]
+
+        def test_tree_value_init_with_item(self):
+            tv1 = treevalue_class({'a': 1, 'b': 2, 'c': {'x': 2, 'y': 3}})
+            assert tv1['a'] == 1
+            assert tv1['b'] == 2
+
+            tv2 = treevalue_class(tv1)
+            assert tv2['a'] == 1
+            assert tv2['b'] == 2
+
+            tv3 = treevalue_class({'a': tv1, 'b': tv2, 'c': tv1})
+            assert tv3['a']['a'] == 1
+            assert tv3['b']['a'] == 1
+            assert tv3['c']['a'] == 1
+
+            with pytest.raises(KeyError):
+                _ = tv3['g']
+            with pytest.raises(TypeError):
+                _ = tv3[0]
 
     return _TestClass
